requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Babel==4.0.0
WTForms==3.1.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
google-generativeai==0.3.2
kagglehub==0.2.2
Pillow==10.1.0
email-validator==2.1.0
gunicorn==21.2.0
